+++ src/types/PromptTemplatePipelineString.ts
@@ -5,7 +5,6 @@
  * It can have 3 formats:
  * -   _(this)_ **.ptbk.md file** in custom markdown format described above
  * -   **JSON** format, parsed from the .ptbk.md file
- * -   **Object** which is created from JSON format and bound with tools around (but not the execution logic)
  *
  * @see https://github.com/webgptorg/promptbook#prompt-template-pipeline
  */